+++ http/default.go
@@ -7,6 +7,7 @@ import (
 
 var (
 	defaultShutdownTimeout = 30 * time.Second
+	defaultErrCode         = "ERR_UNKNOWN"
 )
 
 func defaultResponseOptions() ResponseOptions {
@@ -19,12 +20,14 @@ func defaultResponseOptions() ResponseOptions {
 
 func defaultErrorOptions() ErrorResponseOptions {
 	return ErrorResponseOptions{
-		ErrCode: "ERR_UNKNOWN",
 		ResponseOptions: ResponseOptions{
 			EncodeFunc:  EncodeJSON,
 			ContentType: ApplicationJSON,
 			CharsetType: UTF8,
 		},
+		Err:     nil,
+		ErrCode: defaultErrCode,
+		ErrData: nil,
 	}
 }